current weather displaying application using API
This is a simple Current weather update model using API which shows some feature of the current weather of the different countries and cities all around the world and it stores all the output current data into a excel sheet. i have used openpyxl to store the data into excel and the various library used in this are Pandas,request and json.